refactor(api-credentials): isolate telemetry adapters
